requirements.txt
pandas
numpy
matplotlib
seaborn 
scikit-learn
xgboost
joblib
ipykernel 
langchain
langchain-groq
groq
python-dotenv